allow transparent faking of RPi.GPIO
