+++ fake_rpi/__init__.py
@@ -1,13 +1,11 @@
 from __future__ import print_function
 from . import serial
 import fake_rpi.smbus
-from . import RPi
+import fake_rpi.RPi
 from .Adafruit import LSM303
 from .wrappers import printf
 from .wrappers import toggle_print
 from . import picamera
-from .RPi import PWM
-from .RPi import GPIO
 from .version import __version__
 
 __author__ = 'Kevin Walchko'
